requirements.txt
Django==4.2
psycopg2-binary
praw
openai
django-mptt